Refactor imports to `lib.get_reference_node` since the other function is deprecated
