+++ openpype/hosts/maya/plugins/inventory/import_reference.py
@@ -1,7 +1,7 @@
 from maya import cmds
 
 from openpype.pipeline import InventoryAction
-from openpype.hosts.maya.api.plugin import get_reference_node
+from openpype.hosts.maya.api.lib import get_reference_node
 
 
 class ImportReference(InventoryAction):
